% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{fetch_pancreas}
\alias{fetch_pancreas}
\title{Fetch pancreas dataset}
\usage{
fetch_pancreas(data_home = NULL, download_if_missing = TRUE)
}
\arguments{
\item{data_home}{Optional path to directory for storing data}

\item{download_if_missing}{Whether to download if data not found locally}
}
\value{
SingleCellExperiment object or NULL if not found and download_if_missing=FALSE
}
\description{
Fetch pancreas dataset
}
