% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaussiancopula.R
\name{GaussianCopula}
\alias{GaussianCopula}
\title{Gaussian Copula Simulator}
\description{
Gaussian Copula Simulator

Gaussian Copula Simulator
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-GaussianCopula-new}{\code{GaussianCopula$new()}}
\item \href{#method-GaussianCopula-fit}{\code{GaussianCopula$fit()}}
\item \href{#method-GaussianCopula-parameters}{\code{GaussianCopula$parameters()}}
\item \href{#method-GaussianCopula-sample}{\code{GaussianCopula$sample()}}
\item \href{#method-GaussianCopula-stop}{\code{GaussianCopula$stop()}}
\item \href{#method-GaussianCopula-clone}{\code{GaussianCopula$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GaussianCopula-new"></a>}}
\if{latex}{\out{\hypertarget{method-GaussianCopula-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GaussianCopula$new(
  mean_formula = NULL,
  sdev_formula = NULL,
  copula_formula = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GaussianCopula-fit"></a>}}
\if{latex}{\out{\hypertarget{method-GaussianCopula-fit}{}}}
\subsection{Method \code{fit()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GaussianCopula$fit(sce, max_epochs = 50L, lr = 0.1, batch_size = 1024L)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GaussianCopula-parameters"></a>}}
\if{latex}{\out{\hypertarget{method-GaussianCopula-parameters}{}}}
\subsection{Method \code{parameters()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GaussianCopula$parameters()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GaussianCopula-sample"></a>}}
\if{latex}{\out{\hypertarget{method-GaussianCopula-sample}{}}}
\subsection{Method \code{sample()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GaussianCopula$sample(obs = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GaussianCopula-stop"></a>}}
\if{latex}{\out{\hypertarget{method-GaussianCopula-stop}{}}}
\subsection{Method \code{stop()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GaussianCopula$stop()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GaussianCopula-clone"></a>}}
\if{latex}{\out{\hypertarget{method-GaussianCopula-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GaussianCopula$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
