% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/negbincopula.R
\name{NegBinCopula}
\alias{NegBinCopula}
\title{Negative Binomial Copula Simulator}
\description{
Negative Binomial Copula Simulator

Negative Binomial Copula Simulator
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-NegBinCopula-new}{\code{NegBinCopula$new()}}
\item \href{#method-NegBinCopula-fit}{\code{NegBinCopula$fit()}}
\item \href{#method-NegBinCopula-parameters}{\code{NegBinCopula$parameters()}}
\item \href{#method-NegBinCopula-sample}{\code{NegBinCopula$sample()}}
\item \href{#method-NegBinCopula-stop}{\code{NegBinCopula$stop()}}
\item \href{#method-NegBinCopula-clone}{\code{NegBinCopula$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NegBinCopula-new"></a>}}
\if{latex}{\out{\hypertarget{method-NegBinCopula-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NegBinCopula$new(
  mean_formula = NULL,
  dispersion_formula = NULL,
  copula_formula = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NegBinCopula-fit"></a>}}
\if{latex}{\out{\hypertarget{method-NegBinCopula-fit}{}}}
\subsection{Method \code{fit()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NegBinCopula$fit(sce, max_epochs = 50L, lr = 0.1, batch_size = 1024L)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NegBinCopula-parameters"></a>}}
\if{latex}{\out{\hypertarget{method-NegBinCopula-parameters}{}}}
\subsection{Method \code{parameters()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NegBinCopula$parameters()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NegBinCopula-sample"></a>}}
\if{latex}{\out{\hypertarget{method-NegBinCopula-sample}{}}}
\subsection{Method \code{sample()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NegBinCopula$sample(obs = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NegBinCopula-stop"></a>}}
\if{latex}{\out{\hypertarget{method-NegBinCopula-stop}{}}}
\subsection{Method \code{stop()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NegBinCopula$stop()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NegBinCopula-clone"></a>}}
\if{latex}{\out{\hypertarget{method-NegBinCopula-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NegBinCopula$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
